pyproject.toml
[project]
name = "lakeflow"
version = "0.1.0"
description = "Lakeflow MCP"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.81.0",
    "mcp>=1.26.0",
    "typer>=0.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lakeflow.py"]
